requirements.txt
Flask==2.0.1
Flask-MySQL==1.5.1